+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ##### 2.0 (not yet released)
 
+* an error message is shown when attempting to create a new map with the same name as already existing one
+
 ##### 1.1 (26/01/2020)
 
 * added image icons for markers